+++ index.html
@@ -99,6 +99,10 @@
         </div>
    
     </div>
+    <!-- COPY URL POP UP -->
+    <div id="url-copy-container" class="url-copy-container">
+      <span>URL Copied to Clipboard</span>
+  </div>
   </main>
 
   <script type="module" src="main.js">
