+++ src/components/FontSizeMenu.js
@@ -1,0 +1,17 @@
+export function FontSizeMenu () {
+  return (
+         ` 
+         <div id="font-size-menu" class="drop-down-menu-container">
+              <ul className="dropdown-menu" aria-labelledby="font-size-menu">
+                  <li data-id="font-size" >14 px</li>
+                  <li data-id="font-size" >15 px</li>
+                  <li data-id="font-size" >16 px</li>
+                  <li data-id="font-size" >17 px</li>
+                  <li data-id="font-size" >18 px</li>
+                  <li data-id="font-size" >19 px</li>
+                  <li data-id="font-size" >20 px</li>
+              </ul>
+          </div> 
+          `
+  )
+}
