+++ src/components/ThemeMenu.js
@@ -1,0 +1,13 @@
+export function ThemeMenu () {
+  return (
+           ` 
+            <div id="theme-menu" class="drop-down-menu-container">
+                <ul className="dropdown-menu" aria-labelledby="theme-menu">
+                    <li data-id="theme" >vs-dark</li>
+                    <li data-id="theme" >hc-black</li>
+                    <li data-id="theme" >hc-light</li>
+                </ul>
+            </div> 
+            `
+  )
+}
