--- src/editorOptions.js
@@ -1,18 +1,0 @@
-export const editorOptions = {
-  value: '',
-  language: 'html',
-  theme: 'vs-dark',
-  automaticLayout: true,
-  renderLineHighlight: 'none',
-  fontFamily: 'Cascadia Code',
-  fontSize: '15px',
-  overviewRulerBorder: false,
-  cursorSmoothCaretAnimation: 'on',
-  cursorBlinking: 'expand',
-  cursorWidth: 3,
-  overviewRulerLanes: 0,
-  minimap: { enabled: false },
-  padding: { top: '20px' },
-  codeLens: true
-
-}
