+++ src/utils/API.js
@@ -1,0 +1,5 @@
+// CONSTANTS
+export const API = {
+  LINK_SHORTENER_URL: 'https://simple-api-server-dev-xrtr.3.us-1.fl0.io/cody/'
+
+}
